Extract type to perform clustering operation
